VSFlow.ConstantPanel — Type

ConstantPanel

Structure representing a vortex panel of uniform strength.

VSFlow.LinearPanel — Type

LinearPanel

Structure representing a vortex panel of linearily varying strength.

VSFlow.Profile — Type

PROFILE object

VSFlow.Profile — Method

Constructor Profile(profile, N, position, args...)

VSFlow.VortexPoint — Type

VortexPoint

Structure to represent a point vortex.

VSFlow.BuildPanel — Method

BuildPanel(T, X1, Y1, X2, Y2, profID)

Build a panel of type T starting at point X1, Y1 and ending at point X2, Y2.

VSFlow.backwarddifference — Method

backwarddifference(x, old_x, dt)

Compute the backward difference of nth order according to the length of old_x.

VSFlow.blobkernel — Method

blobkernel(X, Y, vpX, vpY, δ)

Return the velocity at point X, Y induced by a vortex point located at vpX,vpY.

VSFlow.centereddifference — Method

centereddifference(neighbours, dx)

Centered differences. Order depends on length of neighbours.

VSFlow.checkangle — Method

checkangle(θ)

Assert the continuity of panel angles by subtracting 2π if necessary.

VSFlow.circularmotion — Method

circularmotion(t, R, strouhal)

Return the vector [x, ẋ, ẍ] corresponding to a circular motion of radius R where x = [X, Y, α].

VSFlow.correctvelocity — Method

correctvelocity(xm, ym, xb, yb, xt, yt, ub, vb, ut, vt)

Compute the angular velocity (ratio between angular impulse and scalar inertia moment). Integrals computed with bi-linear assumption on quadrilaterals, and 25-point Gauss quadrature.

VSFlow.cropedge! — Method

cropegde!(p::Panel, ϵ, isend)

Shortens the panel by a length ϵ at its end if isend == true or at its start otherwise.

VSFlow.cubicspline — Method

cubicspline(z, X, U)

Computes the cubic spline at point z ∈ [0, 1], build from interpolation of points (X, U) as well as its derivative.

VSFlow.customfourthorder — Method

customfourthorder(x, old_x, dt)

Compute the 4th order time derivative at t using t-1/2 t-3/2 t-5/2 t-7/2.

VSFlow.dipolecombgenerator — Function

dipolecombgenerator(rng::MersenneTwister, T, δ, threshold=1e-2)

VSFlow.gausslegendretriangle — Method

gausslegendretriangle(f, fx, fy)

Computes the integral of function f(x,y) on the standard triangle abc: a (0; 0) b (1; 0) c (0; 1) using 25 points per dimension. fx, fy are the mapping from the standard triangle to the real one. Coefficients are extracted from Rathod et al (2004) "Gauss Legendre quadrature over a triangle"

VSFlow.genprofile — Method

genprofile(profile, N, position, args...)

Generate a profile with a wedged trailing edge.

VSFlow.getangularimpulsecorrection — Method

getangularimpulsecorrection(p::LinearPanel, ω, Up, Vp, Ut, Vt, X0, Y0, Xp, Yp)

Return the angular impulse correction at panel p to obtain the moment with a compact formulation depending on velocity and its derivatives. This correction is necessary if integrals are computed in the body reference frame.

VSFlow.getbodycentroid — Method

getbodycentroid(p::LinearPanel, X0, Y0)

Returns the contribution of panel p on the position of the centroid of the body with respect to position X0, Y0, multiplied by the body volume.

VSFlow.getbodymass — Method

getbodymass(XYm, XYt, XYb)

Integrate the constant and uniform unit density on the body. Integral computed with 25-point Gauss quadrature.

VSFlow.getbodymomentofinertia — Method

getbodymomentofinertia(p::LinearPanel, X0, Y0)

Return the contribution of panel p on the moment of inertia of the body.

VSFlow.getbodyvolume — Method

getbodyvolume(p::LinearPanel, X0, Y0)

Return the contribution of panel p on the body volume.

VSFlow.getboundpanelsinducedvel — Method

getboundpanelsinducedvel(profile, X, Y)

VSFlow.getboundpotential — Function

getboundpotential(p::Profile, nn=20) Integrates the velocity on the profile to obtain the potential. Integrating on a path from upstream is required to get the exact pressure. If the goal is to get the loads, then the upstream integration is unnecessary.

VSFlow.getcenterofmass — Method

getcenterofmass(xm, ym, xb, yb, xt, yt)

Compute the center of mass of the body. Integrals computed with 25-point Gauss quadrature.

VSFlow.getcenterofmassvelocity — Method

getcenterofmassvelocity(xm, ym, xb, yb, xt, yt, ub, vb, ut, vt)

Compute the velocity of the center of mass of the body. Integrals computed with bi-linear assumption on quadrilaterals.

VSFlow.getcentroid — Method

getcentroid(p::LinearPanel, γL, γR)

Return the absolute position (in reference frame) of the vorticity centroid of panel p.

VSFlow.getcloudvelocities — Method

getcloudvelocities(vortex_points, δ)

Return the velocity induced by all other vortices in the cloud.

VSFlow.getcoeffimpulse! — Function

getcoeffimpulse!()

VSFlow.getcoeffpotential! — Method

getcoeffpotential!

VSFlow.getconstantspeedcoeffs — Method

getconstantspeedcoeffs(p::LinearPanel, o::ConstantPanel, vertical)

Return the velocity parameters induced by uniform panel o at center of panel p. If vertical == true, then only the normal components are returned.

VSFlow.getenergyflux — Function

getenergyflux(p::LinearPanel, γL, γR, ω, Ut, Vt, X0, Y0, Xp, Yp, vs=0)

Return the flux of energy at panel p to obtain the force/moment with a compact formulation depending on velocity and its derivatives.

VSFlow.getforceerror! — Method

getforceerror!(p::Profile, sindex, tindex)

VSFlow.getimpulse — Function

getimpulse(p::LinearPanel, γL, γR, ω, Ut, Vt, X0, Y0, Xp, Yp, nc, vs=0)

Compute the impulse by integrating X^γ and X^X^γ analytically on the linear panel. You can specify if you need to add the integral of X^(n^Vb) where n points towards the fluid. The integral is performed in the panel frame and the result is then rotated in the inertial frame of reference. Inertial frame: X, Panel frame: -x + x'

Input: -p : the linear panel to integrate on -γL : vortex strength on the left side of the panel (in its frame) -γR : vortex strength on the right side of the panel -ω : angular velocity in the z direction -Ut : translational horizontal body velocity -Vt : translational vertical body velocity -X0 : position X of origin of the inertial frame -Y0 : position Y of origin of the inertial frame -Xp : position X of the center of rotation of the body -Yp : position Y of the center of rotation of the body -nc : need body velocity contribution? -vs : stretching velocity

VSFlow.getimpulse — Method

getimpulse(p::Profile)

VSFlow.getinducedvelocity — Function

getinducedvelocity(p::LinearPanel, XY, k, γL, γR, ω, Ut, Vt, Xp, Yp, vs)

Computes the velocity due to panel p at location XY k specifies which integral is required: 0 full integral; vortex 1 full integral; source 2 half-left integral; vortex 3 half-left integral; source 4 half-right integral; vortex 5 half-right integral; source

VSFlow.getinducedvelocity — Method

getinducedvelocity(vp::VortexPoint, X, Y, δ)

Return the velocity intuced by vortex vp at location X, Y.

VSFlow.getlinearspeedcoeffs — Method

getlinearspeedcoeffs(p::LinearPanel, o::LinearPanel, vertical)

Return the velocity parameters induced by linear panel o at center of panel p. If vertical == true, then only the normal components are returned.

VSFlow.getmonopolesource — Method

getmonopolesource(p::LinearPanel, id, ω, Ut, Vt, Xp, Yp)

Compute the monopole of source of a linear panel and its location in the reference frame. If vortex strength crosses the x axis, then the panel is split into to smaller panels of positive and negative strengths. In the case it is split, then the id is modified to be able to retrace it later.

Input: -p : the linear panel to integrate on -id : the id of the panel

VSFlow.getmonopolevortex — Function

getmonopolevortex(p::LinearPanel, id, γL, γR, ω, Ut, Vt, Xp, Yp, vs=0)

Compute the monopole of vorticity of a linear panel and its location in the reference frame. If vortex strength crosses the x axis, then the panel is split into to smaller panels of positive and negative strengths. In the case it is split, then the id is modified to be able to retrace it later.

Input: -p : the linear panel to integrate on -id : the id of the panel -γL : vortex strength on the left side of the panel (in its frame) -γR : vortex strength on the right side of the panel

VSFlow.getmotionkelvin — Method

getmotionkelvin(p::LinearPanel, ω, Ut, Vt, X0, Y0)

Get the circulation induced by the motion of panel p.

Input: -p : the linear panel to integrate on -ω : angular velocity in the z direction -Ut : translational horizontal body velocity -Vt : translational vertical body velocity -X0 : position X of the center of rotation of the body -Y0 : position Y of the center of rotation of the body

VSFlow.getmotionparams — Function

getmotionparams(p::LinearPanel, X, Y)

Get the velocity parameters at point X, Y due to the motion of panel p. Integrals are performed in the panel frame of reference then re-cast in the inertial frame through a rotation.

Input: -p : the linear panel to integrate on -X : position X where we want the velocity induced by the body motion -Y : position Y where we want the velocity induced by the body motion -ω : angular velocity in the z direction -Ut : translational horizontal body velocity -Vt : translational vertical body velocity -X0 : position X of the center of rotation of the body -Y0 : position Y of the center of rotation of the body -vs : stretching velocity

VSFlow.getnoca! — Function

VSFlow.getnormal — Method

getnormal(p::LinearPanel)

Get the normal of the panel in its frame of reference (inside of the body).

VSFlow.getparams — Method

getparams(p::ConstantPanel, X, Y)

Return the velocity induced by panel p if it had unit strength at location X, Y.

VSFlow.getparams — Method

getparams(p::LinearPanel, X, Y)

Obtain the parameters to multiply with γL, γR in order to get the velocity at point X, Y induced by linear vortex panel p.

VSFlow.getparamsinrefframe — Method

getparamsinrefframe(panel, X, Y)

VSFlow.getparamsinrefframe — Method

getparamsinrefframe(panel, X, Y)

VSFlow.getpotential — Method

getpotential(p::LinearPanel, γL, γR, ω, Ut, Xp, Yp)

Compute the potential through the integral of tangent velocity along the panel p.

VSFlow.getpressureforce — Method

getpressureforce(p::LinearPanel, p1, p2, Xp, Yp)

Return the force on panel p due to pressure distribution.

VSFlow.getprofilelengths — Method

getprofilelengths(p::Profile)

VSFlow.getstate — Method

getstate(p, n) Used for the gym model

VSFlow.gettrailingedgevel! — Method

gettrailingedgevel!(p::Profile)

VSFlow.getvelocitygradient — Method

getvelocitygradient(vp::VortexPoint, X, Y, θ)

Computes the gradient of induced velocity at position X, Y due to a change of position of vp.

Input : -vp : sensitive point vortex -X : X position where velocity is computed -Y : Y position where velocity is computed -θ : rotation angle -δ : Gaussian kernel cut-off width

VSFlow.getvorteximpulse — Method

getvorteximpulse(p, vp)

VSFlow.getγ — Method

getγ(p::LinearPanel, γ1, γ2, X, Y)

Return the vortex strength of panel p at point X, Y. This point must be part of the panel.

VSFlow.get∫fdV — Method

get∫fdV(XY, f)

Computes the integral of f on the triangle created by the set of points XY, Y using a Gauss-Legendre quadrature. f can return a vector.

VSFlow.get∫fdV — Method

get∫fdV(pa::LinearPanel, XY, f)

Computes the integral of f on the triangle created by the panel and point XY using a Gauss-Legendre quadrature. f can return a vector.

VSFlow.globalvelocity — Function

globalvelocity(p::Profile, XY)

VSFlow.heavepitch — Method

heavepitch(t, h0, αmax, ψ1, ψ2, strouhal)

Return the vector [x, ẋ, ẍ] corresponding to a heaving & pitching motion where x = [X, Y, α].

VSFlow.inbodyframe — Method

inbodyframe(p::Profile, X, Y)

VSFlow.inpanelframe — Method

inpanelframe(p::Panel, X, Y)

Return the point X, Y in the frame of panel p.

VSFlow.inreferenceframe — Method

inreferenceframe(p::Panel, X, Y)

Return the point x, y in the reference frame.

VSFlow.logresults — Function

logresults(p, is_init, t, mem)

VSFlow.logresults2 — Function

logresults2(p, is_init, t, mem) Results for the RL controller

VSFlow.lumpvortices! — Method

lumpvortices!(p::Profile)

VSFlow.motionupdate! — Method

motionupdate!(p::Profile, X, Ẋ, t)

VSFlow.norm2 — Method

norm2(x)

Squared norm of x.

VSFlow.placeconstantpanel! — Method

placeconstantpanel!(p::Profile, X2, Y2)

VSFlow.profilerun — Function

VSFlow.resetconstantpanel! — Method

function resetconstantpanel!(p) Transform the constant panel into a point vortex. Then recompute the vortex strength distribution on the body.

VSFlow.rotate — Method

rotate(X, Y, θ)

Rotate frame by angle θ in the clockwise direction.

VSFlow.setangle! — Method

setangle!(p::Panel, α, pivotlocation)

Rotate panel p by an angle α with rotation center located at pivotlocation.

VSFlow.setcps! — Function

setcps

VSFlow.setforcecontrol! — Method

setforcecontrol!(p::Profile, fx, fy, mz)

VSFlow.setinitvelocity — Method

setinitvelocity(p::Profile, u, v, α̇)

VSFlow.setγs! — Method

setγs!(p::Profile)

VSFlow.step! — Function

step!(p::Profile)

VSFlow.transfer! — Method

transfer!(p, sindex, tindex)

VSFlow.translate! — Method

translate!(p::Panel, Xt, Yt)

Translate panel p.

VSFlow.trapz — Method

trapz(f, x)

Integrate function f using trapeze rule.