Public API

Airfoil geometries

Airfoil motion

Running the simulation

Post processing

Private API

Vortex Elements

Airfoil geometries

Maths

Public API

Airfoil geometries

VSFlow.circle — Function

circle(R)(x)

Return the y coordinates of a circle of radius R.

VSFlow.ellipse — Function

ellipse(B)(x)

Return the y coordinates of an ellipse with half height B.

VSFlow.gaw1 — Function

gaw1(x)

Return the y coordinates of a GAW1 airfoil.

VSFlow.naca00 — Function

naca00(ZZ)(x)

Return the y coordinates of a NACA00ZZ airfoil.

Airfoil motion

VSFlow.uniform — Function

uniform(vx, vy, α̇)(t)

Return the vector [x, ẋ, ẍ] corresponding to a uniform straight motion in the [-X, Y, -θ] direction with velocity [vx, vy, α̇].

VSFlow.heavepitch — Function

heavepitch(h0, αmax, ψ1, ψ2, strouhal)(t)

Return the vector [x, ẋ, ẍ] corresponding to a heaving & pitching motion where x = [-X, Y, α].

VSFlow.circularmotion — Function

circularmotion(R, strouhal)(t)

Return the vector [x, ẋ, ẍ] corresponding to a circular motion of radius R where x = [X, Y, α].

Running the simulation

VSFlow.Profile — Method

Profile(; id, profileshape::Function, x0, ẋ0, N, dt, T, δ = 1e-2, ϵ = 1e-2, (η, Tmin, Smin) = zeros(3))

Constructs a Profile object. If eta, Tmin, Smin are all 0, then no lumping operation occurs.

Keyword Arguments

id: String used to identify the profile.

profileshape: function determining the profile shape.

N: number of panels on the surface of the profile.

x0: initial position [X0, Y0, α0] of the quarter chord of the profile.

ẋ0: initial velocity [Ẋ0, Ẏ0, α̇0] of the quarter chord of the profile.

dt: timestep of the simulation.

T: horizon time.

δ = 1e-2: kernel cut-off width.

ϵ = 1e-2: percentage of the average panel length to chop the trailing edge.

η = 0: maximum error due to lumping.

Tmin = 0: minimum time between two successive active vortices.

Smin = 0: minimum vortex sheet length in the wake.

VSFlow.profilerun — Function

profilerun(p::Profile, accfunc, isshow=false; is4thorder=true)

Simulate the testcase.

Arguments

p: Profile simulation to run.

accfunc: Function giving the force applied on the pivot point of the body.

isshow = false: Bool indicating whether an animation has to be generated.

Keyword Arguments

is4thorder = true: Bool indicating if the time marching is RK4 or ForwardEuler.

Post processing

VSFlow.History — Type

History

A structure allowing to record useful data.

Fields

t: timelapse.

X: position [X, Y, α].

Ẋ: velocity [U, V, ω].

acp: aerodynamic coefficients obtained with pressure integration.

acn: aerodynamic coefficients obtained with control volume approach.

aci: aerodynamic coefficients obtained with momentum conservation.

ϕs: velocity potential on body surface.

cps: pressure coefficients on body surface.

Γ: total body circulation.

θ: shedding angle (0 is the bissector).

VSFlow.getimpulse — Method

getimpulse(p::Profile)

Return the linear/angular impulse exerted by the fluid on the body.

VSFlow.getboundpotential — Function

getboundpotential(p::Profile)

Integrates the velocity on the profile to obtain the velocity potential ϕ.

VSFlow.getcoeffimpulse! — Function

getcoeffimpulse!(p::Profile, update=true, β=1.)

Return aerodynamic coefficients computed with impulse conservation in the flow.

Keyword Arguments

update = true: indicates whether the impulse history has to be updated.

β = 1.0: portion of a timestep.

VSFlow.getcoeffpotential! — Function

getcoeffpotential!(p::Profile, update = true)

Return the aerodynamic coefficients through pressure integration on the body.

Keyword Arguments

update = true: indicates whether the impulse history has to be updated.

VSFlow.getnoca! — Function

getnoca!(p::Profile, update=true)

Return aerodynamic coefficients computed with a control volume approach.

Keyword Arguments

update = true: indicates whether the impulse history has to be updated.

Private API

Vortex Elements

VSFlow.VortexPoint — Type

VortexPoint

Structure representing a point vortex.

Fields

X: X-position of the point vortex in the inertial frame.

Y: Y-position of the point vortex in the inertial frame.

Γ: Circulation of the point vortex.

dv: Velocity correction in case of lumping.

VSFlow.ConstantPanel — Type

ConstantPanel

Structure representing a vortex panel of uniform strength.

Fields

X1: X-position of the start of the panel in the inertial frame.

Y1: Y-position of the start of the panel in the inertial frame.

X2: X-position of the end of the panel in the inertial frame.

Y2: Y-position of the end of the panel in the inertial frame.

Xc: X-position of the center of the panel in the inertial frame.

Yc: Y-position of the center of the panel in the inertial frame.

b: half-length of the panel.

θ: panel orientation with respect to the X axis of the inertial frame.

profID: ID of the profile the panel belongs to.

VSFlow.LinearPanel — Type

LinearPanel

Structure representing a vortex panel of linearly varying strength.

Fields

X1: X-position of the start of the panel in the inertial frame.

Y1: Y-position of the start of the panel in the inertial frame.

X2: X-position of the end of the panel in the inertial frame.

Y2: Y-position of the end of the panel in the inertial frame.

Xc: X-position of the center of the panel in the inertial frame.

Yc: Y-position of the center of the panel in the inertial frame.

b: half-length of the panel.

θ: panel orientation with respect to the X axis of the inertial frame.

profID: ID of the profile the panel belongs to.

Airfoil geometries

VSFlow.initiator — Function

initiator(ξ, T)

Smooth increase from 0 to 1 in half a period T and its time derivative.

VSFlow.carlingfish — Function

carlingfish(s, t, T)

Return the curvature of a curve of unit length associated with the motion of a carling fish of period T. Its time derivative is computed too.

VSFlow.splinefish — Function

splinefish(s, t, T, X, K)

Return the curvature of a curve of unit length associated with the motion of a fish following a cubic spline.

VSFlow.swimmerthickness — Function

swimmerthickness(s)

Return the thickness of the swimming body along its coordinate s.

VSFlow.getbodymass — Function

getbodymass(XYm, XYt, XYb)

Integrate the constant and uniform unit density on the body. Integral computed with 25-point Gauss quadrature.

VSFlow.getinertia — Function

getinertia(XYm, XYt, XYb)

Return the moment of inertia of the body through 25-point Gauss quatradure on triangles.

VSFlow.getcenterofmass — Function

getcenterofmass(xm, ym, xb, yb, xt, yt)

Compute the center of mass of the body. Integrals computed with 25-point Gauss quadrature.

VSFlow.getcenterofmassvelocity — Function

getcenterofmassvelocity(xm, ym, xb, yb, xt, yt, ub, vb, ut, vt)

Compute the velocity of the center of mass of the body. Integrals computed with bi-linear assumption on quadrilaterals.

VSFlow.correctvelocity — Function

correctvelocity(xm, ym, xb, yb, xt, yt, ub, vb, ut, vt)

Compute the angular velocity (ratio between angular impulse and scalar inertia moment). Integrals computed with bi-linear assumption on quadrilaterals, and 25-point Gauss quadrature.

Maths

VSFlow.rotate — Function

rotate(X, Y, θ)

Rotate frame by angle θ in the clockwise direction.

VSFlow.simps — Function

simps(f, x)

Integrate function f using Simpson's rule.

VSFlow.trapz — Function

trapz(f, x)

Integrate function f using trapeze rule.

VSFlow.backwarddifference — Function

backwarddifference(x, old_x, dt)

Compute the backward difference of nth order according to the length of old_x.

VSFlow.centereddifference — Function

centereddifference(neighbours, dx)

Centered differences. Order depends on length of neighbours.

VSFlow.checkangle — Function

checkangle(θ)

Assert the continuity of panel angles by subtracting 2π if necessary.

VSFlow.gausslegendretriangle — Function

gausslegendretriangle(f, fx, fy)

Computes the integral of function f(x,y) on the standard triangle abc: a (0; 0) b (1; 0) c (0; 1) using 25 points per dimension. fx, fy are the mapping from the standard triangle to the real one. Coefficients are extracted from Rathod et al (2004) "Gauss Legendre quadrature over a triangle"

VSFlow.get∫fdV — Function

get∫fdV(XY, f)

Computes the integral of f on the triangle created by the set of points XY, Y using a Gauss-Legendre quadrature. f can return a vector.

get∫fdV(pa::LinearPanel, XY, f)

Computes the integral of f on the triangle created by the panel and point XY using a Gauss-Legendre quadrature. f can return a vector.

VSFlow.cubicspline — Function

cubicspline(z, X, U)

Computes the cubic spline at point z ∈ [0, 1], build from interpolation of points (X, U) as well as its derivative.