VSFlow.wgauss — Constant

Constant values for Gauss quadrature in triangles

VSFlow.Profile — Type

PROFILE object

VSFlow.Profile — Method

Constructor Profile(profile, N, position, args...)

VSFlow.backwarddifference — Method

Backward differences Compute the backward difference of nth order.

VSFlow.centereddifference — Method

Centered differences

VSFlow.checkangle — Method

Assert the continuity of panel angles

VSFlow.correctvelocity — Method

function correctvelocity(xm, ym, xb, yb, xt, yt, ut, vt, ub, vb) Compute the angular velocity (ratio between angular impulse and scalar inertia moment). Integrals computed with bi-linear assumption on quadrilaterals, and 25-point Gauss quadrature.

VSFlow.cubicspline — Method

cubicspline(z, X, U) Computes the cubic spline at point z ∈ [0, 1], build from interpolation of points (X, U) as well as its derivative.

VSFlow.customfourthorder — Method

function customfourthorder(x, old_x, dx) Compute the 4th order time derivative at t using t-1/2 t-3/2 t-5/2 t-7/2.

VSFlow.dipolecombgenerator — Function

dipolecombgenerator(rng::MersenneTwister, T, δ, threshold=1e-2)

VSFlow.gausslegendretriangle — Method

function gausslegendretriangle(f, x, y) Computes the integral of function f(x,y) on the standard triangle abc: a (0; 0) b (1; 0) c (0; 1) using 25 points per dimension. fx, fy are the mapping from the standard triangle to the real one. Coefficients are extracted from Rathod et al (2004) "Gauss Legendre quadrature over a triangle"

VSFlow.genprofile — Method

genprofile(profile, N, position, args...)

Generate a profile with a wedged trailing edge.

VSFlow.getbodycentroid — Method

function getbodycentroid(p::LinearPanel, X0, Y0) returns the contribution of panel "p" on the position of the centroid of the body with respect to position "X0, Y0", multiplied by the body volume.

VSFlow.getbodymass — Method

function getbodymass(XYm, XYt, XYb) Integrate the constant and uniform unit density on the body (fish). Integral computed with 25-point Gauss quadrature.

VSFlow.getbodymomentofinertia — Method

function getbodymomentofinertia(p::LinearPanel, X0, Y0) returns the contribution of panel "p" on the moment of inertia of the body.

VSFlow.getbodyvolume — Method

function getbodyvolume(p::LinearPanel, X0, Y0) returns the contribution of panel "p" on the body volume.

VSFlow.getboundpanelsinducedvel — Method

getboundpanelsinducedvel(profile, X, Y)

VSFlow.getboundpotential — Function

getboundpotential(p::Profile, nn=20) Integrates the velocity on the profile to obtain the potential. Integrating on a path from upstream is required to get the exact pressure. If the goal is to get the loads, then the upstream integration is unnecessary.

VSFlow.getcenterofmass — Method

function getcenterofmass(xm, ym, xb, yb, xt, yt) Compute the center of mass of the body (fish). Integrals computed with 25-point Gauss quadrature.

VSFlow.getcenterofmassvelocity — Method

function getcenterofmassvelocity(xm, ym, xb, yb, xt, yt) Compute the velocity of the center of mass of the body (fish). Integrals computed with bi-linear assumption on quadrilaterals.

VSFlow.getcentroid — Method

function getcentroid(p::LinearPanel) returns the absolute position of the vorticity centroid of the panel.

VSFlow.getcloudvelocities — Method

vortexcloudvelocity This function uses anti-symmetry to obtain the velocities induced by the vortex cloud on itself

VSFlow.getcoeffimpulse! — Function

getcoeffimpulse!()

VSFlow.getcoeffpotential! — Method

getcoeffpotential!

VSFlow.getforceerror! — Method

getforceerror!(p::Profile, sindex, tindex)

VSFlow.getimpulse — Function

function getimpulse(p::LinearPanel, γ1, γ2) Compute the impulse by integrating X^γ and X^X^γ analytically on the linear panel. You can specify if you need to add the integral of X^(n^Vb) where n points towards the fluid. The integral is performed in the panel frame and the result is then rotated in the inertial frame of reference. Inertial frame: X Panel frame: -x + x'

Input:
	-p  : the linear panel to integrate on
	-γL : vortex strength on the left side of the panel (in its frame)
	-γR : vortex strength on the right side of the panel
	-ω  : angular velocity in the z direction
	-Ut : translational horizontal body velocity
	-Vt : translational vertical body velocity
	-X0 : position X of origin of the inertial frame
	-Y0 : position Y of origin of the inertial frame
	-Xp : position X of the center of rotation of the body
	-Yp : position Y of the center of rotation of the body
	-nc : need body velocity contribution?
	-vs : stretching velocity

VSFlow.getimpulse — Method

getimpulse(p::Profile)

VSFlow.getinducedvelocity — Function

function getinducedvelocity(p::LinearPanel, XY, k, γL, γR, ω, Ut, Vt, Xp, Yp, vs) Computes the velocity due to pane p at location XY k specifies which integral is required: 0 full integral; vortex 1 full integral; source 2 half-left integral; vortex 3 half-left integral; source 4 half-right integral; vortex 5 half-right integral; source

VSFlow.getmonopolesource — Method

function getmonopolesource() Compute the monopole of source of a linear panel and its location in the reference frame. If vortex strength crosses the x axis, then the panel is split into to smaller panels of positive and negative strengths. In the case it is split, then the id is modified to be able to retrace it later.

Input:
	-p  : the linear panel to integrate on
    -id : the id of the panel

VSFlow.getmonopolevortex — Function

function getmonopolevortex() Compute the monopole of vorticity of a linear panel and its location in the reference frame. If vortex strength crosses the x axis, then the panel is split into to smaller panels of positive and negative strengths. In the case it is split, then the id is modified to be able to retrace it later.

Input:
	-p  : the linear panel to integrate on
    -id : the id of the panel
	-γL : vortex strength on the left side of the panel (in its frame)
	-γR : vortex strength on the right side of the panel

VSFlow.getmotionkelvin — Method

function getmotionkelvin(p::LinearPanel, ω, Ut, Vt, X0, Y0) Get the circulation induced by the panel motion.

Input:
	-p  : the linear panel to integrate on
	-ω  : angular velocity in the z direction
	-Ut : translational horizontal body velocity
	-Vt : translational vertical body velocity
	-X0 : position X of the center of rotation of the body
	-Y0 : position Y of the center of rotation of the body

VSFlow.getmotionparams — Function

function getmotionparams(p::LinearPanel, X, Y) Get the velocity parameters at point X, Y due to the motion of the panel. Integrals are performed in the panel frame of reference then re-cast in the inertial frame through a rotation.

Input:
	-p  : the linear panel to integrate on
	-X  : position X where we want the velocity induced by the body motion
	-Y  : position Y where we want the velocity induced by the body motion
	-ω  : angular velocity in the z direction
	-Ut : translational horizontal body velocity
	-Vt : translational vertical body velocity
	-X0 : position X of the center of rotation of the body
	-Y0 : position Y of the center of rotation of the body
	-vs : stretching velocity

VSFlow.getnoca! — Function

VSFlow.getnormal — Method

function getnormal(p::LinearPanel) Get the normal of the panel in its frame of reference (inside of the body)

VSFlow.getparams — Method

function getparams(p::LinearPanel, X, Y) Obtain the parameters to multiply with γL, γR in order to get the velocity at point X, Y.

VSFlow.getparamsinrefframe — Method

getparamsinrefframe(panel, X, Y)

VSFlow.getparamsinrefframe — Method

getparamsinrefframe(panel, X, Y)

VSFlow.getpotential — Method

function getpotential(p::LinearPanel, γL, γR, ω, Ut, Xp, Yp) Compute the potential through the integral of tangent velocity along the panel "p".

VSFlow.getpressureforce — Method

function getpressureforce(p::LinearPanel)

VSFlow.getprofilelengths — Method

getprofilelengths(p::Profile)

VSFlow.getstate — Method

getstate(p, n) Used for the gym model

VSFlow.gettrailingedgevel! — Method

gettrailingedgevel!(p::Profile)

VSFlow.getvelocitygradient — Method

getvelocitygradient(vp::VortexPoint, X, Y, θ) Computes the gradient of induced velocity at position 'X, Y' due to a change of position of 'vp'

Input :
    -vp : sensitive point vortex
    -X  : X position where velocity is computed
    -Y  : Y position where velocity is computed
    -θ  : rotation angle
    -δ  : Gaussian kernel cut-off width

VSFlow.getvorteximpulse — Method

getvorteximpulse(p, vp)

VSFlow.get∫fdV — Method

function get∫fdV(XY, f) Computes the integral of 'f' on the triangle created by the set of points XYs, Ys using a Gauss-Legendre quadrature. f can return a vector.

VSFlow.get∫fdV — Method

function get∫fdV(pa::LinearPanel, XY, f) Computes the integral of 'f' on the triangle created by the panel and point XY using a Gauss-Legendre quadrature. f can return a vector.

VSFlow.globalvelocity — Function

globalvelocity(p::Profile, XY)

VSFlow.heavepitch — Method

Heave & Pitch foil

VSFlow.inbodyframe — Method

inbodyframe(p::Profile, X, Y)

VSFlow.logresults — Function

logresults(p, is_init, t, mem)

VSFlow.logresults2 — Function

logresults2(p, is_init, t, mem) Results for the RL controller

VSFlow.lumpvortices! — Method

lumpvortices!(p::Profile)

VSFlow.motionupdate! — Method

motionupdate!(p::Profile, X, Ẋ, t)

VSFlow.norm2 — Method

Squared norm of "x".

VSFlow.placeconstantpanel! — Method

placeconstantpanel!(p::Profile, X2, Y2)

VSFlow.profilerun — Function

VSFlow.resetconstantpanel! — Method

function resetconstantpanel!(p) Transform the constant panel into a point vortex. Then recompute the vortex strength distribution on the body.

VSFlow.rotate — Method

Rotate frame by angle "θ" in the clockwise direction.

VSFlow.setcps! — Function

setcps

VSFlow.setforcecontrol! — Method

setforcecontrol!(p::Profile, fx, fy, mz)

VSFlow.setinitvelocity — Method

setinitvelocity(p::Profile, u, v, α̇)

VSFlow.setγs! — Method

setγs!(p::Profile)

VSFlow.step! — Function

step!(p::Profile)

VSFlow.transfer! — Method

transfer!(p, sindex, tindex)

VSFlow.trapz — Method

Integrate using trapeze rule.