Public API

Airfoil geometries

VSFlow.naca00 — Function

naca00(x; kwargs...)

Return the coordinates of a NACA00ZZ airfoil.

Keyword arguments:

ZZ = 12: max thickness (in % of the chord) of the airfoil.

VSFlow.gaw1 — Function

gaw1(x)

Return the coordinates of a GAW1 airfoil.

VSFlow.circle — Function

circle(x; kwargs...)

Return the coordinates of a circle.

Keyword arguments:

R = .5: radius of the circle.

VSFlow.ellipse — Function

ellipse(x; kwargs...)

Return the coordinates of an ellipse.

Keyword arguments:

B = .5: half height of the ellipse.

Airfoil motion

VSFlow.heavepitch — Function

heavepitch(t, h0, αmax, ψ1, ψ2, strouhal)

Return the vector [x, ẋ, ẍ] corresponding to a heaving & pitching motion where x = [X, Y, α].

VSFlow.circularmotion — Function

circularmotion(t, R, strouhal)

Return the vector [x, ẋ, ẍ] corresponding to a circular motion of radius R where x = [X, Y, α].

Running the simulation

VSFlow.Profile — Method

Profile(id, profile::Function, N, position, dt, T, δ, ϵ, rng, is_comb, lump, args...)

Constructs a Profile object.

Arguments

-id: String used to identify the profile. -profile: function determining the profile shape. -N: number of panels on the surface of the profile. -position: position of the pivot point of the profile. -dt: timestep of the simulation. -T: horizon time. -δ: kernel cut-off width. -ϵ: percentage of the average panel length to chop the trailing edge. -rng: random number generator -is_comb: Bool indicating whether a random comb of dipoles hit the body. -lump: Bool indicating whether the lumging of vortices should occur.

Additional Arguments

-η: maximum error due to lumping. -Tmin: minimum time between two successive active vortices. -sheet_size: minimum vortex sheet length in the wake. -ZZ: body maximum thickness (in % of the chord).

VSFlow.setinitvelocity — Function

setinitvelocity(p::Profile, u, v, α̇)

Set the initial velocities of the body. Note: α̇ represents the rate of change in AoA and is thus opposite to the angular velocity.

VSFlow.profilerun — Function

profilerun(p::Profile, is_write, is_lumped, is4thorder, forcefun, motion_args, show=false)

Simulate the testcase.

Arguments

-p: -is_write: Bool indicating if the results are saved into a file. -is_lumped: Bool indicating if vortices have to be lumped together. -is4thorder: Bool indicating if the time marching is RK4 or ForwardEuler. -forcefun: function giving the force applied on the pivot point of the body. -motion_args: array containing parameters to feed forcefun with. -show: Bool indicating whether an animation has to be generated.

Post Processing

Private API

Vortex Elements

Airfoil geometries

VSFlow.initiator — Function

initiator(ξ, T)

Smooth increase from 0 to 1 in half a period T and its time derivative.

VSFlow.carlingfish — Function

carlingfish(s, t, T)

Return the curvature of a curve of unit length associated with the motion of a carling fish of period T. Its time derivative is computed too.

VSFlow.splinefish — Function

splinefish(s, t, T, X, K)

Return the curvature of a curve of unit length associated with the motion of a fish following a cubic spline.

VSFlow.swimmerthickness — Function

swimmerthickness(s)

Return the thickness of the swimming body along its coordinate s.

VSFlow.getbodymass — Function

getbodymass(XYm, XYt, XYb)

Integrate the constant and uniform unit density on the body. Integral computed with 25-point Gauss quadrature.

VSFlow.getinertia — Function

getinertia(XYm, XYt, XYb)

Return the moment of inertia of the body through 25-point Gauss quatradure on triangles.

VSFlow.getcenterofmass — Function

getcenterofmass(xm, ym, xb, yb, xt, yt)

Compute the center of mass of the body. Integrals computed with 25-point Gauss quadrature.

VSFlow.getcenterofmassvelocity — Function

getcenterofmassvelocity(xm, ym, xb, yb, xt, yt, ub, vb, ut, vt)

Compute the velocity of the center of mass of the body. Integrals computed with bi-linear assumption on quadrilaterals.

VSFlow.correctvelocity — Function

correctvelocity(xm, ym, xb, yb, xt, yt, ub, vb, ut, vt)

Compute the angular velocity (ratio between angular impulse and scalar inertia moment). Integrals computed with bi-linear assumption on quadrilaterals, and 25-point Gauss quadrature.

Maths

VSFlow.rotate — Function

rotate(X, Y, θ)

Rotate frame by angle θ in the clockwise direction.

Missing docstring.

Missing docstring for VSFlow.simps. Check Documenter's build log for details.

VSFlow.trapz — Function

trapz(f, x)

Integrate function f using trapeze rule.

VSFlow.backwarddifference — Function

backwarddifference(x, old_x, dt)

Compute the backward difference of nth order according to the length of old_x.

VSFlow.centereddifference — Function

centereddifference(neighbours, dx)

Centered differences. Order depends on length of neighbours.

VSFlow.checkangle — Function

checkangle(θ)

Assert the continuity of panel angles by subtracting 2π if necessary.

VSFlow.gausslegendretriangle — Function

gausslegendretriangle(f, fx, fy)

Computes the integral of function f(x,y) on the standard triangle abc: a (0; 0) b (1; 0) c (0; 1) using 25 points per dimension. fx, fy are the mapping from the standard triangle to the real one. Coefficients are extracted from Rathod et al (2004) "Gauss Legendre quadrature over a triangle"

VSFlow.get∫fdV — Function

get∫fdV(XY, f)

Computes the integral of f on the triangle created by the set of points XY, Y using a Gauss-Legendre quadrature. f can return a vector.

get∫fdV(pa::LinearPanel, XY, f)

Computes the integral of f on the triangle created by the panel and point XY using a Gauss-Legendre quadrature. f can return a vector.

VSFlow.cubicspline — Function

cubicspline(z, X, U)

Computes the cubic spline at point z ∈ [0, 1], build from interpolation of points (X, U) as well as its derivative.